@@ -38,9 +38,12 @@
     "@react-three/drei": "^9.65.4",
     "@react-three/fiber": "^8.12.1",
     "@redux-devtools/extension": "^3.2.5",
+    "@solana/wallet-adapter-base": "^0.9.23",
     "@solana/wallet-adapter-nextjs-starter": "^0.8.17",
-    "@solana/web3.js": "^1.78.4",
-    "@spling/social-protocol": "^1.5.9",
+    "@solana/wallet-adapter-react": "^0.15.35",
+    "@solana/wallet-adapter-react-ui": "^0.9.33",
+    "@solana/wallet-adapter-wallets": "^0.19.25",
+    "@solana/web3.js": "^1.89.1",
     "@types/react": "^18.0.32",
     "@types/react-dom": "18.0.11",
     "@types/three": "^0.155.0",
@@ -53,21 +56,20 @@
     "bs58": "^5.0.0",
     "dotenv": "^16.3.1",
     "eslint": "^8.37.0",
-    "eslint-config-next": "13.2.4",
-    "fluent-ffmpeg": "^2.1.2",
     "formidable": "^3.5.1",
     "framer-motion": "^10.12.4",
     "fs": "^0.0.1-security",
     "maplibre-gl": "^3.5.1",
     "micro": "^10.0.1",
     "mongodb": "^5.3.0",
-    "next": "^12.3.4",
-    "next-auth": "^4.24.5",
     "party-js": "^2.2.0",
     "passport-google-oidc": "^0.1.0",
+    "next": "^14.0.4",
+    "node-fetch": "^3.3.2",
+    "pino-pretty": "^10.2.3",
     "querystring": "^0.2.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hot-toast": "^2.4.1",
     "react-map-gl": "^7.1.6",
     "react-use": "^17.4.0",
@@ -83,6 +85,7 @@
     "autoprefixer": "^10.4.14",
     "babel-loader": "^9.1.3",
     "daisyui": "^3.6.3",
+    "eslint-config-next": "^14.0.4",
     "eslint-config-prettier": "^8.8.0",
     "postcss": "^8.4.21",
     "prettier": "^2.8.7",
