@@ -1,7 +1,7 @@
 import React from "react";
 import styles from "../layout.module.css";
-import Image from "next/image";
-import { useRouter } from "next/router";
+import Image from "next/legacy/image";
+import { useRouter } from "next/navigation";
 
 interface HomeBtnProps {
   handleClick: () => void;
@@ -13,16 +13,6 @@ export const HomeBtn: React.FC<HomeBtnProps> = ({ handleClick }) => {
   return (
     <>
       <button onClick={handleClick} className={`${styles["app-title"]}`}>
-        {/* <h1
-          className={`self-start text-2xl break-all absolute italic w-5 ${styles["app-title"]}`}
-        >
-          øhr
-        </h1>
-        <h1
-          className={`self-start text-2xl break-all absolute italic w-5 ${styles["app-title-top"]}`}
-        >
-          øhr
-        </h1> */}
         <Image
           src={"/three_d_logo.png"}
           alt="'øhr' written in 3D"
